requirements.txt
pandas
numpy
requests
python-dotenv
vaderSentiment
torch
transformers
google-generativeai
matplotlib
seaborn
ruptures